pyproject.toml
[build-system]
requires = ["setuptools>=61.2.0", "setuptools_scm[toml]>=3.4.3"]
build-backend = "setuptools.build_meta"

[project]
name = "drs-controller"
authors = [
    {name = "Mirantis OSCORE team"},
]
description = "Dynamic Resource (re)Scheduler"
readme = "README.rst"
requires-python = ">=3.10"
keywords = ["Kubernetes", "OpenStack"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "kopf",
    "openstacksdk",
    "pykube-ng",
    "prometheus-api-client",
    "stevedore",
]
dynamic = ["version"]

[tool.setuptools]
packages = ["drs_controller"]

[tool.setuptools_scm]

[project.optional-dependencies]
test = ["black", "ruff"]

[project.entry-points."drs_controller.collector"]
stacklight = "drs_controller.collectors.stacklight:collector"

[project.entry-points."drs_controller.scheduler"]
ease-most-loaded = "drs_controller.schedulers.ease_most_loaded:scheduler"

[project.entry-points."drs_controller.mover"]
os-live-migration = "drs_controller.movers.openstack_live_migration:mover"
